tilgangsstyring/requirements.txt
snowbird @ git+https://github.com/navikt/snowbird@v0.5
streamlit
snowflake-connector-python[pandas]
snowflake-cli-labs
watchdog
dbt-snowflake<1.9.0,>=1.8.0
dbt-core<1.11.0,>=1.10.0